// Back_Office/Front/src/models/module-model.ts
export enum ModuleEnum {
  Dashboard = "dashboard",
  Roles = "roles",
  Users = "users",
  Audits = "audits",
  Settings = "settings",
  Degrees = "degrees",
  Technologies = "technologies",
  Profiles = "profiles",
  Durations = "durations",
  Types = "types",
  Subjects = "subjects",
  Candidatures = "candidatures",
  EmailTemplates = "email_templates",
  FrontOfficeMessages = "front_office_messages",
  EmailLogs = "email_logs",
  MailConfig = "mail_config",
}